Move private dps data classes out of cognite.client.data_classes.datapoints
